refactor(storage): extract row mappers into mapper.go (SRP) (#201)
Move all SQL row-scanning and struct-building functions out of queries.go
into a new mapper.go file as package-level functions. DB methods now handle
only query construction and execution; mapper.go owns data mapping.

Closes #201
